requirements.txt
# Vercel-optimized requirements
# Core dependencies only - heavy packages removed for serverless

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Instagram API (lightweight)
instagrapi==2.1.2
requests==2.31.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pytz==2024.1

# Security
cryptography==42.0.0
pyjwt==2.8.0
bcrypt==4.1.2
